Completed Code. Treat as Beta. Passed End-to-End test.
Added Jump-to-App. Add an improved test app (App2). Bootloader run if,
within the 20 second start window, it sees it's address OR after the 20
sec, it finds no App to jump to. Apps are assumed to be C code so the
entry point is 000F (word address after Intr Jump Table).