ButtonLED project added
The project uses the ATmega4809 Curiosity Nano Evaluation kit to demonstrate a
simple project framework. It uses the following support files:
   sysTime     - timer function to provide millis().
   mod_led     - general LED control.
I include these in just about every project I make.

The demo monitors the on-board button and flashes the on-board LED twice when the
button is pushed.
If the button is held down, the LED flashes three times two seconds after the
initial double flash.